% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{build_mesh}
\alias{build_mesh}
\alias{build_mesh,Domain,numeric,numeric-method}
\title{Delaunay triangulation of the spatial domain}
\usage{
build_mesh(domain, maximum_area, minimum_angle)

\S4method{build_mesh}{Domain,numeric,numeric}(domain, maximum_area, minimum_angle)
}
\arguments{
\item{domain}{a domain object returned by \code{Domain}.}

\item{maximum_area}{maximum triangle area.}

\item{minimum_angle}{minimum triangle angle in degrees.}
}
\description{
Delaunay triangulation of the spatial domain
}
