% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{contour,Function-method}
\alias{contour,Function-method}
\title{Create a contour plot of a Function}
\usage{
\S4method{contour}{Function}(x, ...)
}
\arguments{
\item{x}{A \code{Function} generated by \code{Function}}

\item{...}{Arguments representing graphical options to be passed to \code{\link[plotly]{plot_ly}}.}
}
\value{
A plotly object
}
\description{
Create a contour plot of a Function
}
