% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{contour,FunctionObject-method}
\alias{contour,FunctionObject-method}
\title{Create a contour plot of a FunctionObject}
\usage{
\S4method{contour}{FunctionObject}(x, ...)
}
\arguments{
\item{x}{A \code{FunctionObject} generated by \code{Function}}

\item{...}{Arguments representing graphical options to be passed to plot_ly function.}
}
\value{
A plotly object
}
\description{
Create a contour plot of a FunctionObject
}
