% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{DifferentialOps}
\alias{DifferentialOps}
\alias{+.DiffOpObject}
\alias{*.DiffOpObject}
\title{plus operator overload for DiffOpObject}
\usage{
\method{+}{DiffOpObject}(e1, e2)

\method{*}{DiffOpObject}(e1, e2)
}
\arguments{
\item{e1}{a numeric.}

\item{e2}{a DiffOpObject.}
}
\value{
A S4 object representing the sum of two differential operators.

A S4 object representing the product by scalar for a differential operator.
}
\description{
plus operator overload for DiffOpObject

product by scalar for differential operators
}
