% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{laplace}
\alias{laplace}
\alias{div}
\title{laplace operator for Function class}
\usage{
laplace(Function)

div(Function)
}
\arguments{
\item{Function}{a Function.}
}
\value{
A R6 class representing the laplace operator applied to the function passed as parameter.

A R6 class representing the diffusion term of a second order linear differential operator.
}
\description{
laplace operator for Function class

divergence operator FunctionGrad
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
laplace_f <- laplace(f)
}
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
K <- matrix(c(1,2,1,0),nrow=2,ncol=2)
diffusion <- div(K*grad(f))
} 
}
