% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{div}
\alias{div}
\title{divergence operator FunctionGradObject}
\usage{
div(f)
}
\arguments{
\item{f}{a FunctionObject.}
}
\value{
A S4 object representing the diffusion term of a second order linear differential operator.
}
\description{
divergence operator FunctionGradObject
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
K <- matrix(c(1,2,1,0),nrow=2,ncol=2)
diffusion <- div(K*grad(f))
} 
}
