% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{Domain}
\alias{Domain}
\alias{Domain,list-method}
\alias{Domain,pslg-method}
\alias{Domain,sfc-method}
\title{R6 class representing a spatial (spatio-temporal) domain}
\usage{
Domain(x)

\S4method{Domain}{list}(x)

\S4method{Domain}{pslg}(x)

\S4method{Domain}{sfc}(x)
}
\arguments{
\item{x}{could be a \code{pslg} returned by \code{\link[RTriangle]{triangulate}}, 
a \code{sfc} returned by \code{\link[sf]{st_geometry}} or a named list containing:
\itemize{
   \item{\code{nodes}, a #nodes-by-2 matrix containing the x and y coordinates of the boundary nodes;}
}}
}
\description{
R6 class representing a spatial (spatio-temporal) domain
}
