% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R, R/mesh.R
\name{\%X\%}
\alias{\%X\%}
\alias{\%X\%,Domain,numeric-method}
\alias{\%X\%,Mesh,numeric-method}
\title{create spatio-temporal domain}
\usage{
op1 \%X\% op2

\S4method{\%X\%}{Domain,numeric}(op1, op2)

\S4method{\%X\%}{Mesh,numeric}(op1, op2)
}
\arguments{
\item{op1}{A mesh object created by \code{Mesh}.}

\item{op2}{A numeric vector.}
}
\value{
An S4 object representing a spatio-temporal domain.
}
\description{
create spatio-temporal domain
}
