% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{-.DifferentialOp}
\alias{-.DifferentialOp}
\title{Difference of differential operators}
\usage{
\method{-}{DifferentialOp}(e1, e2)
}
\arguments{
\item{e1}{a \code{DifferentialOp}.}

\item{e2}{a \code{DifferentialOp} or missing.}
}
\value{
A R6 class representing the difference of two differential operators.
}
\description{
Difference of differential operators
}
