% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{dot}
\alias{dot}
\alias{dot,vector,FunctionGradObject-method}
\title{dot product}
\usage{
dot(op1, op2)

\S4method{dot}{vector,FunctionGradObject}(op1, op2)
}
\arguments{
\item{op1}{a numeric vector.}

\item{op2}{a FunctionGradObject.}
}
\value{
A S4 object representing the advection term of a second order linear differential operator.
}
\description{
dot product
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
b <- c(1,1)
advection <- dot(b,grad(f))
} 
}
