\name{femR-package}
\alias{femR-package}
\alias{femR}
\docType{package}
\title{
\packageTitle{femR}
}
\description{
\packageDescription{femR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{femR}
\packageIndices{femR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{femR}

Maintainer: \packageMaintainer{femR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
