% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{Function}
\alias{Function}
\title{Finite Element Function}
\usage{
Function(FunctionSpace)
}
\arguments{
\item{FunctionSpace}{a FunctionSpace object created by \code{FunctionSpace}:}
}
\value{
An S4 object representing a Function belonging to the FunctionSpace passed as parameter.
}
\description{
Finite Element Function

Create Function object
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
}
}
