% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{laplace}
\alias{laplace}
\title{laplace operator for Function}
\usage{
laplace(f)
}
\arguments{
\item{f}{a Function.}
}
\value{
A S4 object representing the laplace operator applied to the function passed as parameter.
}
\description{
laplace operator for Function
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
laplace_f <- laplace(f)
}
}
