% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{Mesh}
\alias{Mesh}
\alias{Mesh,list-method}
\alias{Mesh,triangulation-method}
\title{Create mesh object}
\usage{
Mesh(domain)

\S4method{Mesh}{list}(domain)

\S4method{Mesh}{triangulation}(domain)
}
\arguments{
\item{domain}{A named list containing must contain:
\itemize{
   \item{\code{nodes}, a #nodes-by-2 matrix containing the x and y coordinates of the mesh nodes;}
   \item{\code{elements}, a #elements-by-3 matrix specifiying the triangles giving the row's indices in \code{nodes} of the triangles' vertices;}
   \item{\code{boundary}, a #nodes-by-1 matrix, with entries either '1' or '0'. An entry '1' indicates that the corresponding node is a boundary node; 
          an entry '0' indicates that the corresponding node is not a boundary node.}
}}
}
\value{
An S4 object representing a Mesh.
}
\description{
Create mesh object
}
