% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{-,DiffOpObject,DiffOpObject-method}
\alias{-,DiffOpObject,DiffOpObject-method}
\alias{-,DiffOpObject,missing-method}
\title{difference of differential operators}
\usage{
\S4method{-}{DiffOpObject,DiffOpObject}(e1, e2)

\S4method{-}{DiffOpObject,missing}(e1)
}
\arguments{
\item{e1}{a DiffOpObject.}

\item{e2}{a DiffOpObject.}
}
\value{
A S4 object representing the difference of two differential operators.

A S4 object representing the sum of two differential operators.
}
\description{
difference of differential operators

minus (unary) operator for DiffOpObject
}
