% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{DifferentialOps}
\alias{DifferentialOps}
\alias{-.DiffOpObject}
\title{difference of differential operators}
\usage{
\method{-}{DiffOpObject}(e1, e2)
}
\arguments{
\item{e1}{a DiffOpObject.}

\item{e2}{a DiffOpObject.}
}
\value{
A S4 object representing the difference of two differential operators.
}
\description{
difference of differential operators
}
