% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde.R
\name{Pde}
\alias{Pde}
\alias{Pde,DifferentialOp,function-method}
\alias{Pde,DifferentialOp,numeric-method}
\title{A PDEs object}
\usage{
Pde(DifferentialOp, forcing)

\S4method{Pde}{DifferentialOp,function}(DifferentialOp, forcing)

\S4method{Pde}{DifferentialOp,numeric}(DifferentialOp, forcing)
}
\arguments{
\item{DifferentialOp}{a differential operator.}

\item{forcing}{a standard R function representing the forcing term of the PDE or a numeric value, in case of constant forcing term.}
}
\value{
A R6 class representing a PDE.
}
\description{
A PDEs object
}
