% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{plot,FunctionObject-method}
\alias{plot,FunctionObject-method}
\title{Plot a Function object}
\usage{
\S4method{plot}{FunctionObject}(x, y, ...)
}
\arguments{
\item{x}{A \code{FunctionObject} generated by \code{Function}}

\item{...}{Arguments representing graphical options to be passed to plot_ly function.}
}
\value{
A plotly object
}
\description{
Plot a Function object
}
