% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{plot.FunctionObject}
\alias{plot.FunctionObject}
\title{Plot a Function object}
\usage{
\method{plot}{FunctionObject}(x, ...)
}
\arguments{
\item{x}{A \code{FunctionObject} generated by \code{Function}}

\item{...}{Arguments representing graphical options to be passed to \code{\link[plotly]{plot_ly}}.}
}
\value{
A plotly object
}
\description{
Plot a Function object
}
