% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{plot.MeshObject}
\alias{plot.MeshObject}
\title{Plot a Mesh object}
\usage{
\method{plot}{MeshObject}(x, ...)
}
\arguments{
\item{x}{A \code{MeshObject} object defining the triangular mesh, as generated by \code{Mesh}}

\item{...}{Arguments representing graphical options to be passed to \code{\link[plotly]{plot_ly}}.}
}
\value{
A plotly object
}
\description{
Plot a Mesh object
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
plot(mesh)
}
}
