% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.DifferentialOp}
\alias{+.DifferentialOp}
\title{Sum of DifferentialOp}
\usage{
\method{+}{DifferentialOp}(e1, e2)
}
\arguments{
\item{e1}{a \code{DifferentialOp}.}

\item{e2}{a \code{DifferentialOp}.}
}
\value{
A R6 class representing the sum of two differential operators.
}
\description{
Sum of DifferentialOp
}
