% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+,DiffOpObject,DiffOpObject-method}
\alias{+,DiffOpObject,DiffOpObject-method}
\title{plus operator overload for DiffOpObject}
\usage{
\S4method{+}{DiffOpObject,DiffOpObject}(e1, e2)
}
\arguments{
\item{e1}{a DiffOpObject.}

\item{e2}{a DiffOpObject.}
}
\value{
A S4 object representing the sum of two differential operators.
}
\description{
plus operator overload for DiffOpObject
}
