% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{ReactionOperator}
\alias{ReactionOperator}
\alias{*.Function}
\title{Reaction Operator}
\usage{
\method{*}{Function}(e1, e2)
}
\arguments{
\item{e1}{a numeric.}

\item{e2}{a FunctioObject created by \code{Function}.}
}
\value{
A S4 object representing the reaction term of a second order linear differential operator.
}
\description{
Reaction Operator

product by scalar for FunctionObejct
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
reaction <- 2*f
}
}
