% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mesh.R
\name{read_mesh}
\alias{read_mesh}
\title{reads mesh file from FreeFem++}
\usage{
read_mesh(filename)
}
\arguments{
\item{filename}{path to a .mesh file created by FreeFem++.}
}
\value{
a named list to be passed to \code{Mesh}.
}
\description{
reads mesh file from FreeFem++
}
