% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{st_as_sf.Domain}
\alias{st_as_sf.Domain}
\alias{st_as_sfc.Domain}
\alias{st_as_sfc.Mesh}
\alias{st_crs.Domain}
\alias{st_crs<-.Domain}
\title{sf methods for Domain and Mesh}
\usage{
\method{st_as_sf}{Domain}(x, ...)

\method{st_as_sfc}{Domain}(x, ...)

\method{st_as_sfc}{Mesh}(x, ...)

\method{st_crs}{Domain}(x, ...)

\method{st_crs}{Domain}(x) <- value
}
\arguments{
\item{x}{An object of class \code{Domain} or \code{Mesh}.}

\item{...}{Arguments passed on the corresponding \code{sf} function.}

\item{value}{The value to be assigned. See the documentation of the
corresponding sf function for details.}
}
\value{
The \code{femR} method for \code{\link[sf]{st_as_sf}} returns
............ \code{\link[sf]{sf}}.
}
\description{
\code{\link[sf]{sf}} methods for \code{Domain} and \code{Mesh} objects.
}
