% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{st_as_sf.DomainObject}
\alias{st_as_sf.DomainObject}
\title{sf methods for DomainObject and MeshObject}
\usage{
st_as_sf.DomainObject(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{DomainObject}} or \code{\link{MeshObject}}.}

\item{...}{Arguments passed on the corresponding \code{sf} function.}

\item{y}{An object of class of \code{\link[sf:st]{sfg}}.}

\item{value}{The value to be assigned. See the documentation of the
corresponding sf function for details.}
}
\value{
The \code{femR} method for \code{\link[sf]{st_as_sf}} returns
............ \code{\link[sf]{sf}}.
}
\description{
\code{\link[sf]{sf}} methods for \code{\link{DomainObject}} and \code{\link{MeshObject}} objects.
}
\details{
See the \code{\link[sf]{sf}} documentation.
}
