% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{*.FunctionGrad}
\alias{*.FunctionGrad}
\title{product overload for FunctionGradObejct}
\usage{
\method{*}{FunctionGrad}(e1, e2)
}
\arguments{
\item{e1}{a numeric matrix.}

\item{e2}{a FunctionGrad class created by \code{grad} function.}
}
\value{
A FunctionGrad class.
}
\description{
product overload for FunctionGradObejct
}
