% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{*,matrix,FunctionGrad-method}
\alias{*,matrix,FunctionGrad-method}
\title{product overload for FunctionGradObejct}
\usage{
\S4method{*}{matrix,FunctionGrad}(e1, e2)
}
\arguments{
\item{e1}{a numeric matrix.}

\item{e2}{a FunctionGrad created by \code{grad} function.}
}
\value{
A FunctionGrad.
}
\description{
product overload for FunctionGradObejct
}
