% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{*,numeric,DiffOpObject-method}
\alias{*,numeric,DiffOpObject-method}
\title{product by scalar for differential operators}
\usage{
\S4method{*}{numeric,DiffOpObject}(e1, e2)
}
\arguments{
\item{e1}{a numeric.}

\item{e2}{a DiffOpObject.}
}
\value{
A S4 object representing the product by scalar for a differential operator.
}
\description{
product by scalar for differential operators
}
