% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{times_DifferentialOps}
\alias{times_DifferentialOps}
\alias{*.DifferentialOp}
\title{product by scalar for differential operators}
\usage{
\method{*}{DifferentialOp}(e1, e2)
}
\arguments{
\item{e1}{a numeric.}

\item{e2}{a DifferentialOp.}
}
\value{
A R6 class representing the product by scalar for a differential operator.
}
\description{
product by scalar for differential operators
}
