% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{Triangulate}
\alias{Triangulate}
\alias{Triangulate,DomainObject,numeric,numeric-method}
\title{Delaunay triangulation of the spatial domain}
\usage{
Triangulate(Domain, maximum_area, minimum_angle)

\S4method{Triangulate}{DomainObject,numeric,numeric}(Domain, maximum_area, minimum_angle)
}
\arguments{
\item{Domain}{}

\item{maximum_area}{maximum triangle area}

\item{minumum_angle}{minimum triangle angle in degrees}
}
\description{
Delaunay triangulation of the spatial domain
}
