% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{unit_square}
\alias{unit_square}
\title{Unit square domain}
\description{
The boundary and interior nodes and connectivity matrix of a triangular mesh of
the unit square domain. 
This dataset can be used to create a mesh object with the function \code{Mesh}.
}
