% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{dt,FunctionObject,missing,missing-method}
\alias{dt,FunctionObject,missing,missing-method}
\title{dt function overload for FunctionObejct}
\usage{
\S4method{dt}{FunctionObject,missing,missing}(x, df, ncp)
}
\arguments{
\item{x}{a FunctionObject:}
}
\value{
A S4 object representing the time derivative of a FunctionObject.
}
\description{
dt function overload for FunctionObejct
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
dt(f)
}
}
