% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_space.R
\name{FunctionSpace}
\alias{FunctionSpace}
\alias{FunctionSpace,Mesh,numeric-method}
\alias{FunctionSpace,Mesh,missing-method}
\title{Create FunctionSpace object}
\usage{
FunctionSpace(mesh, fe_order)

\S4method{FunctionSpace}{Mesh,numeric}(mesh, fe_order)

\S4method{FunctionSpace}{Mesh,missing}(mesh)
}
\arguments{
\item{mesh}{A mesh object created by \code{Mesh}:}

\item{fe_order}{Either '1' or '2'. It specifies the finite element order.}
}
\value{
A R6 object representing a Function Space.
}
\description{
Create FunctionSpace object
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh = mesh, fe_order = 1)
}
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
}
}
