% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{grad}
\alias{grad}
\alias{grad,Function-method}
\title{compute gradient of Function}
\usage{
grad(f)

\S4method{grad}{Function}(f)
}
\arguments{
\item{f}{a Function created by \code{Function}:}
}
\value{
An S4 object representing the gradient of the Function passed as parameter.
}
\description{
compute gradient of Function
}
\examples{
\dontrun{
library(femR)
data("unit_square")
mesh <- Mesh(unit_square)
Vh <- FunctionSpace(mesh)
f <- Function(Vh)
grad_f <- grad(f)
}
}
