% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde.R
\name{Pde}
\alias{Pde}
\alias{Pde,DiffOpObject-method}
\title{A PDEs object}
\usage{
Pde(L, u)

\S4method{Pde}{DiffOpObject}(L, u)
}
\arguments{
\item{L}{a differential operator.}

\item{u}{a standard R function representing the forcing term of the PDE.}
}
\value{
A S4 object representing a PDE.
}
\description{
A PDEs object
}
