% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde.R
\name{pde}
\alias{pde}
\alias{Pde}
\alias{Pde,DiffOpObject,function-method}
\alias{Pde,DiffOpObject,numeric-method}
\title{A PDEs object}
\usage{
Pde(DifferentialOp, f)

\S4method{Pde}{DiffOpObject,function}(DifferentialOp, f)

\S4method{Pde}{DiffOpObject,numeric}(DifferentialOp, f)
}
\arguments{
\item{DifferentialOp}{a differential operator.}

\item{f}{a standard R function representing the forcing term of the PDE or a numeric value, in case of constant forcing term.}
}
\value{
A R6 class representing a PDE.
}
\description{
A PDEs object
}
