% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pde.R
\name{Pde}
\alias{Pde}
\alias{Pde,DiffOpObject,ANY-method}
\alias{Pde,DiffOpObject,numeric-method}
\title{A PDEs object}
\usage{
Pde(L, f)

\S4method{Pde}{DiffOpObject,ANY}(L, f)

\S4method{Pde}{DiffOpObject,numeric}(L, f)
}
\arguments{
\item{L}{a differential operator.}

\item{f}{a standard R function representing the forcing term of the PDE or a numeric value, in case of constant forcing term.}
}
\value{
A S4 object representing a PDE.
}
\description{
A PDEs object
}
