% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{st_as_sf.Domain}
\alias{st_as_sf.Domain}
\alias{st_as_sfc.Domain}
\alias{st_as_sfc.Mesh}
\alias{st_crs.Domain}
\alias{st_crs<-.Domain}
\title{Integration with sf package}
\usage{
\method{st_as_sf}{Domain}(x, ...)

\method{st_as_sfc}{Domain}(x, ...)

\method{st_as_sfc}{Mesh}(x, ...)

\method{st_crs}{Domain}(x, ...)

\method{st_crs}{Domain}(x) <- value
}
\arguments{
\item{x}{An object of class \code{Domain} or \code{Mesh}.}

\item{...}{Arguments passed to the corresponding \code{sf} method}

\item{value}{The value to be assigned. See the documentation of the
corresponding sf function for details.}
}
\value{
All methods return the same type of object of the corresponding \code{\link[sf]{sf}} method.
}
\description{
\code{\link[sf]{sf}} methods for \code{Domain} and \code{Mesh} classes.
}
